Fix Crash when not using TC
