+++ src/main/kotlin/rosatech/common/items/RosaMetaItems.kt
@@ -20,7 +20,6 @@ lateinit var INDUSTRIAL_LEATHER_BACKPACK: MetaItem<*>.MetaValueItem
 lateinit var SMART_BACKPACK_LV: MetaItem<*>.MetaValueItem
 lateinit var SMART_BACKPACK_HV: MetaItem<*>.MetaValueItem
 
-lateinit var BIG_ITEM_FILTER: MetaItem<*>.MetaValueItem
 
 class RosaMetaItems {
     companion object {
@@ -39,8 +38,6 @@ class RosaMetaItems {
             SMART_BACKPACK_LV = ITEMS.addItem(3, "smart_backpack.lv").addComponents(SmartBackpackBehavior(63, 1, 4)).setMaxStackSize(1)
             SMART_BACKPACK_HV = ITEMS.addItem(4, "smart_backpack.hv").addComponents(SmartBackpackBehavior(72, 2, 8)).setMaxStackSize(1)
 
-            BIG_ITEM_FILTER = ITEMS.addItem(5, "big_item_filter")
-                .setMaterialInfo(ItemMaterialInfo(MaterialStack(Materials.Zinc, M * 3), MaterialStack(Materials.Aluminium, M)))
         }
     }
 }
